+++ src/useField.ts
@@ -1,14 +1,12 @@
 import * as React from 'react';
 import { formContext } from './helpers/context';
 import { get } from './helpers/operations';
-import reset from './helpers/reset';
 
 export interface FieldOperations {
   onBlur: () => void;
   onChange: (value: any) => void;
   onFocus: () => void;
   setFieldValue: (fieldId: string, value: any) => void;
-  resetField: () => void;
 }
 
 export interface FieldInformation {
@@ -31,21 +29,17 @@ export default function useField(fieldId: string): [FieldOperations, FieldInform
     setFieldTouched,
     touched,
   } = React.useContext(formContext);
-  // Values
-  const error = React.useMemo(() => get(errors, fieldId), [errors]);
-  const initialValue = React.useMemo(() => get(initialValues, fieldId), [initialValues]);
-  const value = React.useMemo(() => get(values, fieldId), [values]);
-  const isFieldTouched = React.useMemo(() => get(touched, fieldId), [touched]);
-  // Methods
-  const resetFieldValue = React.useCallback(() => {
-    setFieldValue(fieldId, initialValue || reset(value));
-    setFieldTouched(fieldId, false);
-  },                                        []);
-  const onChange = React.useCallback((val: any) => setFieldValue(fieldId, val), []);
-  const onBlur = React.useCallback(() => setFieldTouched(fieldId, true), []);
-  const onFocus = React.useCallback(() => setFieldTouched(fieldId, false), []);
   return [
-    { onChange, onBlur, onFocus, resetField: resetFieldValue, setFieldValue },
-    { error, touched: isFieldTouched, value },
+    {
+      onBlur: React.useCallback(() => setFieldTouched(fieldId, true), []),
+      onChange: React.useCallback((val: any) => setFieldValue(fieldId, val), []),
+      onFocus: React.useCallback(() => setFieldTouched(fieldId, false), []),
+      setFieldValue,
+    },
+    {
+      error: React.useMemo(() => get(errors, fieldId), [errors]),
+      touched: React.useMemo(() => get(touched, fieldId), [touched]),
+      value: React.useMemo(() => get(values, fieldId), [values]),
+    },
   ];
 }