+++ src/useFormConnect.ts
@@ -2,8 +2,4 @@ import * as React from 'react';
 import { formContext } from './helpers/context';
 import { FormHookContext } from './types';
 
-const useFormConnect = (): FormHookContext => {
-  return React.useContext(formContext);
-};
-
-export default useFormConnect;
+export default (): FormHookContext => React.useContext(formContext);